feat(framework-issues-ui): Refactor ScannerResultControl (#1355)
* Adding support files for Scenner Result refactoring

* Playing with ScannerResults

* Fixing refactoring errors

* Adding itemSource method

* Making SacannerResultCustomList functional

* Enabling hyperlink click

* Adding middle class to handle shared methods

* Migrating from ScannerResult to customList

* Use a dependency property

* Adding missing shared method

* Addressing PR comments

* Migrating HasUserResizedLvHeader to child

* Get rid of empty parameter comments

* Tying loose wires in ScanerResultCustomListControl

* Cleaning code

* Removing unnecessary properties
